Remove redundant code and tests
