+++ certbot-nginx/certbot_nginx/constants.py
@@ -31,21 +31,6 @@ ALL_SSL_OPTIONS_HASHES = [
 ]
 """SHA256 hashes of the contents of all versions of MOD_SSL_CONF_SRC"""
 
-SSL_DHPARAMS_DEST = "certbot-ssl-dhparams.pem"
-"""Name of the ssl_dhparams file as saved in `IConfig.config_dir`."""
-
-SSL_DHPARAMS_SRC = pkg_resources.resource_filename(
-    "certbot_nginx", "certbot-ssl-dhparams.pem")
-"""Path to the nginx ssl_dhparams file found in the Certbot distribution."""
-
-UPDATED_SSL_DHPARAMS_DIGEST = ".updated-ssl-dhparams-pem-digest.txt"
-"""Name of the hash of the updated or informed ssl_dhparams as saved in `IConfig.config_dir`."""
-
-ALL_SSL_DHPARAMS_HASHES = [
-    '9ba6429597aeed2d8617a7705b56e96d044f64b07971659382e426675105654b',
-]
-"""SHA256 hashes of the contents of all versions of SSL_DHPARAMS_SRC"""
-
 def os_constant(key):
     # XXX TODO: In the future, this could return different constants
     #           based on what OS we are running under.  To see an
