+++ certbot-nginx/certbot_nginx/tests/configurator_test.py
@@ -619,83 +619,6 @@ class InstallSslOptionsConfTest(util.NginxTest):
             "Constants.ALL_SSL_OPTIONS_HASHES must be appended"
             " with the sha256 hash of self.config.mod_ssl_conf when it is updated.")
 
-class InstallSslDhparamsTest(util.NginxTest):
-    """Test that the certbot-ssl-dhparams.pem file is installed and updated properly."""
-
-    def setUp(self):
-        super(InstallSslDhparamsTest, self).setUp()
-
-        self.config = util.get_nginx_configurator(
-            self.config_path, self.config_dir, self.work_dir, self.logs_dir)
-
-    def _call(self):
-        from certbot_nginx.configurator import install_ssl_dhparams
-        install_ssl_dhparams(self.config.ssl_dhparams, self.config.updated_ssl_dhparams_digest)
-
-    def _current_ssl_dhparams_hash(self):
-        from certbot_nginx.constants import SSL_DHPARAMS_SRC
-        return crypto_util.sha256sum(SSL_DHPARAMS_SRC)
-
-    def _assert_current_file(self):
-        self.assertTrue(os.path.isfile(self.config.ssl_dhparams))
-        self.assertEqual(crypto_util.sha256sum(self.config.ssl_dhparams),
-            self._current_ssl_dhparams_hash())
-
-    def test_no_file(self):
-        # prepare should have placed a file there
-        self._assert_current_file()
-        os.remove(self.config.ssl_dhparams)
-        self.assertFalse(os.path.isfile(self.config.ssl_dhparams))
-        self._call()
-        self._assert_current_file()
-
-    def test_current_file(self):
-        self._assert_current_file()
-        self._call()
-        self._assert_current_file()
-
-    def test_prev_file_updates_to_current(self):
-        from certbot_nginx.constants import ALL_SSL_DHPARAMS_HASHES
-        with mock.patch('certbot.crypto_util.sha256sum') as mock_sha256:
-            mock_sha256.return_value = ALL_SSL_DHPARAMS_HASHES[0]
-            self._call()
-        self._assert_current_file()
-
-    def test_manually_modified_current_file_does_not_update(self):
-        with open(self.config.ssl_dhparams, "a") as ssl_dhparams:
-            ssl_dhparams.write("a new line for the wrong hash\n")
-        with mock.patch("certbot.plugins.common.logger") as mock_logger:
-            self._call()
-            self.assertFalse(mock_logger.warning.called)
-        self.assertTrue(os.path.isfile(self.config.ssl_dhparams))
-        self.assertEqual(crypto_util.sha256sum(constants.SSL_DHPARAMS_SRC),
-            self._current_ssl_dhparams_hash())
-        self.assertNotEqual(crypto_util.sha256sum(self.config.ssl_dhparams),
-            self._current_ssl_dhparams_hash())
-
-    def test_manually_modified_past_file_warns(self):
-        with open(self.config.ssl_dhparams, "a") as ssl_dhparams:
-            ssl_dhparams.write("a new line for the wrong hash\n")
-        with open(self.config.updated_ssl_dhparams_digest, "w") as f:
-            f.write("hashofanoldversion")
-        with mock.patch("certbot.plugins.common.logger") as mock_logger:
-            self._call()
-            self.assertEqual(mock_logger.warning.call_args[0][0],
-                "%s has been manually modified; updated file "
-                "saved to %s. We recommend updating %s for security purposes.")
-        self.assertEqual(crypto_util.sha256sum(constants.SSL_DHPARAMS_SRC),
-            self._current_ssl_dhparams_hash())
-        # only print warning once
-        with mock.patch("certbot.plugins.common.logger") as mock_logger:
-            self._call()
-            self.assertFalse(mock_logger.warning.called)
-
-    def test_current_file_hash_in_all_hashes(self):
-        from certbot_nginx.constants import ALL_SSL_DHPARAMS_HASHES
-        self.assertTrue(self._current_ssl_dhparams_hash() in ALL_SSL_DHPARAMS_HASHES,
-            "Constants.ALL_SSL_DHPARAMS_HASHES must be appended"
-            " with the sha256 hash of self.config.ssl_dhparams when it is updated.")
-
 
 if __name__ == "__main__":
     unittest.main()  # pragma: no cover
