+++ certbot/plugins/common_test.py
@@ -169,6 +169,16 @@ class InstallerTest(test_util.ConfigTestCase):
         self.installer.install_ssl_dhparams()
         self.assertTrue(os.path.isfile(self.installer.ssl_dhparams))
 
+    def _current_ssl_dhparams_hash(self):
+        from certbot.constants import SSL_DHPARAMS_SRC
+        return crypto_util.sha256sum(SSL_DHPARAMS_SRC)
+
+    def test_current_file_hash_in_all_hashes(self):
+        from certbot.constants import ALL_SSL_DHPARAMS_HASHES
+        self.assertTrue(self._current_ssl_dhparams_hash() in ALL_SSL_DHPARAMS_HASHES,
+            "Constants.ALL_SSL_DHPARAMS_HASHES must be appended"
+            " with the sha256 hash of self.config.ssl_dhparams when it is updated.")
+
 
 class AddrTest(unittest.TestCase):
     """Tests for certbot.client.plugins.common.Addr."""
